fix: prevent Claude Agent failure from app.isPackaged crash
Agent tasks in the packaged app failed with "Cannot read properties of
undefined (reading 'isPackaged')". The agent worker runs in an Electron
utility process where `electron.app` is undefined, but importing
`htmlToPlainText` from `../../db` pulled in `data-dir.ts` →
`@electron-toolkit/utils`, which evaluates `!electron.app.isPackaged`
at module load.

Move `htmlToPlainText` + `decodeHtmlEntities` to a pure utility module
so the worker's tool files don't transitively load electron/sqlite.
